+++ SolarObjects/ISolarObject.cs
@@ -5,7 +5,7 @@ namespace SolarObjects;
 public interface ISolarObject
 {
     Vector3 Coordinates { get; }
-    int Mass { get; }
+    float Mass { get; }
     void InteractWithAnotherObject(ISolarObject solarObject);
     void Update();
 }
